requirements.txt
numpy
torchvision
matplotlib
tensorboard
terminaltables
pillow
tqdm
imgaug